App now meets specification
User is now presented with a welcome screen if it is their first time using the app. They can then use this to select their currency of choice.

Currencies now have their own object and are declared in the settings file as the currency setting can be changed in different views